major file refactor—thing and other gone, teachers folded into admin folder to reflect that these are admin operations ON teachers, naming of files changed, and AdminTeacherView captures adding and editing teachers entirely. Also, 2 changes—no need to refresh when adding teachers, and edit page redirects back to AdminTeacherView
