+++ app/src/App.tsx
@@ -1,10 +1,8 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Routes, Link } from "react-router-dom";
-import ThingComponent from "./components/ThingComponent";
-import OtherComponent from "./components/Other/OtherComponent";
-import { TeacherProfile } from "./components/Profile/TeacherProfile";
-import { EditTeacherProfile } from "./components/Profile/EditProfile";
-import { EditTeacherPage } from "./components/Profile/EditTeacherPage";
+import { AboutSchool } from "./components/About";
+import { AdminTeacherView } from "./components/Admin/AdminTeacherView";
+import { EditTeacherPage } from "./components/Admin/Teachers/EditTeacherPage";
 import './App.css';
 
 const App: React.FC = () => {
@@ -18,10 +16,7 @@ const App: React.FC = () => {
                 <Link to="/">Home</Link>
               </li>
               <li>
-                <Link to="/allTeachers">View, Remove, Edit All Teachers</Link>
-              </li>
-              <li>
-                <Link to="/buildProfile">Add Teacher</Link>
+                <Link to="/admin/teachers">Admin - View & Change Teacher Roster</Link>
               </li>
             </ul>
           </nav>
@@ -29,10 +24,9 @@ const App: React.FC = () => {
         
 
         <Routes>
-          <Route path="/" element={<ThingComponent />} />
-          <Route path="/allTeachers" element={<OtherComponent />} />
-          <Route path="/buildProfile" element={<TeacherProfile />} />
-          <Route path='/editProfile/:id' element={<EditTeacherPage />} />
+          <Route path="/" element={<AboutSchool />} />
+          <Route path="/admin/teachers" element={<AdminTeacherView />} />
+          <Route path="/editProfile/:id" element={<EditTeacherPage />} />
         </Routes>
       </div>
     </Router>
