+++ app/src/components/About.tsx
@@ -1,0 +1,19 @@
+
+export const AboutSchool: React.FC = () => {
+
+  return (
+    <div>
+      <h1>Home - #MySchool</h1>
+      <h3>All About My School</h3>
+      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, 
+        sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. 
+        Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris 
+        nisi ut aliquip ex ea commodo consequat. 
+        Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore 
+        eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, 
+        sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
+    </div>
+  );
+};
+
+// export default ThingComponent;
